Initial Cleanup
Why these changes are being introduced:
This is a minimalist Python repo for a simple Lambda@Edge function that
is deployed in our CloudFront distribution to re-route https requests
with custom domains. We only need minimal linting/testing.

How this addresses that need:
* Remove the more robust linters and testers
* Simplify the various linting and testing commands

Side effects of this change:
This will not be tested or linted as thoroughly as our other Python
applications, but that is okay for this.