Updates For Caller Workflows
Why these changes are being introduced:
With the new shared workflows now in the main branch of the .github
repository, the final versions of the caller workflow (generated from
the TfC outputs of the mitlib-tf-workloads-libraries-website repo) can
be put in place here.

How this addresses that need:
* Update Makefile with new `create-zip` and `upload-zip` commands
* Update the caller workflows to match the shared workflows in the
.github repository
* Update the README

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/GDT-97