Standard Lambda Handler Naming
Why these changes are being introduced:
There is a tight interdependency between the name of the Python
function and the file in which that function lives and the Terraform
definition of the Lambda that relies on this repository. In order to
keep these in sync, we document the naming decision in both this repo
and in the Terraform repo and make the appropriate changes

See ADR#8 in mitlib-tf-workloads-libraries-website

How this addresses that need:
* Create an ADR in this repo that captures the name of the function that
is the handler function and the name of the .py file that contains that
function.
* Update the function name in the lambda_edge.py file
* In preparation for the full dev/stage/prod workflow, we add the
production deploy GitHub workflow