Fix Production Deploy Workflow
Why these changes are being introduced:
The prod-cf-deploy-lambda.yml workflow did not work correctly on the
first tagged release. Turns out that I had not fully understood what the
GitHub context looked like and the `if:` statement for the job was
incorrect. This addresses that problem.

The GitHub Actions Context is documented at
https://docs.github.com/en/actions/learn-github-actions/contexts. Some
testing in a different repo revealed that there is a key/value pair
that tracks the branch on which the release was tagged:
github.event.release.target_commitish

Later, I discovered that someone on Stack Overflow found this three
years ago:
https://stackoverflow.com/questions/63468238/how-to-get-target-branch-in-github-actions-on-release

Which, after further digging, I discovered was clearly documented in
the GitHub REST API documentation here:
https://docs.github.com/en/rest/releases/releases?apiVersion=2022-11-28

How this addresses that need:
* Correct the `if:` test at the start of the job to correctly identify
if the trigger is a release on `main` or if it's a workflow_dispatch
from the GitHub web UI.

Side effects of this change:
None.